+++ minecraft/common/src/main/java/com/zigythebird/playeranim/accessors/ICapeLayer.java
@@ -1,9 +1,0 @@
-package com.zigythebird.playeranim.accessors;
-
-import com.zigythebird.playeranim.animation.AvatarAnimManager;
-
-@SuppressWarnings("unused")
-public interface ICapeLayer {
-    default void applyBend(AvatarAnimManager manager, float bend) {}
-    default void resetBend() {}
-}